automataCI: added simulation mechanism to test the Release Job API
To make sure the most important and most risky release job, we best
add simulation run capability into it and permits mock run for
testing purposes. That way, it shall not be left out and do crazy
matters anymore.

This patch adds simulation mechanism to test the Release Job API
in automataCI/ directory.

+++ automataCI/_release-docker_windows-any.ps1
@@ -38,18 +38,23 @@ function RELEASE-Run-DOCKER {
 
 	# execute
 	OS-Print-Status info "releasing docker as the latest version..."
-	$__process = DOCKER-Release `
-		"${_target}" `
-		"${_directory}" `
-		"${_datastore}" `
-		"${env:PROJECT_VERSION}"
-	if ($__process -ne 0) {
-		OS-Print-Status error "release failed."
-		return 1
-	}
+	if (-not ([string]::IsNullOrEmpty(${env:PROJECT_SIMULATE_RELEASE_REPO}))) {
+		OS-Print-Status warning "Simulating multiarch docker manifest release..."
+		OS-Print-Status warning "Simulating remove package artifact..."
+	} else {
+		$__process = DOCKER-Release `
+			"${_target}" `
+			"${_directory}" `
+			"${_datastore}" `
+			"${env:PROJECT_VERSION}"
+		if ($__process -ne 0) {
+			OS-Print-Status error "release failed."
+			return 1
+		}
 
-	OS-Print-Status info "remove package artifact..."
-	$null = FS-Remove-Silently "${_target}"
+		OS-Print-Status info "remove package artifact..."
+		$null = FS-Remove-Silently "${_target}"
+	}
 
 	# report status
 	return 0
